pyproject.toml
[tool.poetry]
name = "oz-collabo-006-be"
version = "0.1.0"
description = ""
authors = ["bamjun <khy5116@naver.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
gunicorn = "^20.1.0"
Django = "^5.0.4"
djangorestframework = "^3.15.1"
djangorestframework-simplejwt = "^5.3.1"
psycopg = "^3.1.18"
psycopg2 = "^2.9.9"
django-cors-headers = "^4.3.1"
drf-yasg = "^1.21.7"
boto3 = "^1.34.102"
mysqlclient = "^2.2.4"
selenium = "^4.21.0"
requests = "^2.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

